refactor: changes based on mentor review
