+++ src/components/PokemonSearch/PokemonSearch.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from '../../utils/store';
+import { store } from '../../store/store';
 import {
   mockSuccessfulApiSingleFetch,
   mockFailedFetch,
