+++ src/components/ResultsList/ResultsList.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { store } from '../../utils/store';
+import { store } from '../../store/store';
 import ResultsList from './ResultsList';
 import { mockPokemon } from '../../test-utils/mocks/pokemonapi';
