+++ src/types/pokemonSearch.types.ts
